LPS-135812 Move TemplatePortletKeys to template-api
