+++ modules/apps/template/template-api/bnd.bnd
@@ -1,3 +1,4 @@
 Bundle-Name: Liferay Template API
 Bundle-SymbolicName: com.liferay.template.api
 Bundle-Version: 1.0.0
+Export-Package: com.liferay.template.constants
