+++ modules/apps/template/template-api/src/main/java/com/liferay/template/constants/TemplatePortletKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.template.web.internal.constants;
+package com.liferay.template.constants;
 
 /**
  * @author Lourdes Fernández Besada
