+++ modules/apps/template/template-web/src/main/java/com/liferay/template/web/internal/application/list/TemplatePortletPanelApp.java
@@ -18,7 +18,7 @@ import com.liferay.application.list.BasePanelApp;
 import com.liferay.application.list.PanelApp;
 import com.liferay.application.list.constants.PanelCategoryKeys;
 import com.liferay.portal.kernel.model.Portlet;
-import com.liferay.template.web.internal.constants.TemplatePortletKeys;
+import com.liferay.template.constants.TemplatePortletKeys;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
